+++ hugegraph-client/src/main/java/com/baidu/hugegraph/structure/traverser/KneighborRequest.java
@@ -35,7 +35,7 @@ public class KneighborRequest {
     @JsonProperty("count_only")
     public boolean countOnly = false;
     @JsonProperty("limit")
-    public long limit = Traverser.DEFAULT_LIMIT;
+    public int limit = Traverser.DEFAULT_LIMIT;
     @JsonProperty("with_vertex")
     public boolean withVertex = false;
     @JsonProperty("with_path")
@@ -97,7 +97,7 @@ public class KneighborRequest {
             return this;
         }
 
-        public Builder limit(long limit) {
+        public Builder limit(int limit) {
             TraversersAPI.checkLimit(limit);
             this.request.limit = limit;
             return this;
